+++ src/app/modules/shared/components/create-build-modal/create-build-modal.component.ts
@@ -11,7 +11,6 @@ export class CreateBuildModalComponent {
     constructor(private dialog: MatDialogRef<CreateBuildModalComponent>) {}
     
     public created() {
-        console.log('CreateBuildModalComponent created')
         this.dialog.close();
     }
 }