+++ src/app/modules/shared/components/main-stats/main-stats.component.ts
@@ -185,7 +185,7 @@ export class MainStatsComponent extends AbstractUnsubscribeComponent implements 
         let result = '0';
 
         if (found) {
-            result = this.valueToString(found.total, format.sign, found.suffix);
+            result = this.valueToString(found.totalDisplayed, format.sign, found.suffix);
         }
 
         return result;
@@ -195,6 +195,7 @@ export class MainStatsComponent extends AbstractUnsubscribeComponent implements 
         const found = characterStats.find(characterStat => characterStat.stat === format.stat);
 
         if (found) {
+            console.log(found);
             this.selectedMergedStat = found;
             this.selected.emit(found);
         }